# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Self Hostable Game Library",
        "title": "Rex",
        "contact": {}
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/games/byId": {
            "get": {
                "description": "Get a game by UUID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get game by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the game to search for",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}